Tidy up startVM for all platforms
Factor out the common code to startVM in startvm.go. Keep the
platform-specific details (hypervisors, argument construction) in the
platform-specific getHypervisorArgs function.
