+++ cmd/internal/cli/startvm.go
@@ -6,11 +6,16 @@
 package cli
 
 import (
+	"fmt"
 	"os"
+	"os/exec"
 	"path/filepath"
+	"runtime"
 	"strings"
+	"syscall"
 
 	"github.com/spf13/cobra"
+	"github.com/sylabs/singularity/internal/pkg/buildcfg"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
 )
 
@@ -36,4 +41,51 @@ func prepareVM(cmd *cobra.Command, image string, args []string) {
 		sylog.Errorf("VM instance failed: %s", err)
 		os.Exit(2)
 	}
+}
+
+func startVM(sifImage, singAction, cliExtra string, isInternal bool) error {
+	bzImage := fmt.Sprintf("%s/%s-%s", buildcfg.LIBEXECDIR, "/singularity/vm/syos-kernel", runtime.GOARCH)
+	initramfs := fmt.Sprintf("%s/%s_%s.gz", buildcfg.LIBEXECDIR, "/singularity/vm/initramfs", runtime.GOARCH)
+
+	args := getHypervisorArgs(sifImage, bzImage, initramfs, singAction, cliExtra)
+
+	sylog.Debugf("About to launch VM using: %+v", args)
+
+	hvExec, lookErr := exec.LookPath(args[0])
+	if lookErr != nil {
+		sylog.Fatalf("Failed to find hypervisor executable at %s", args[0])
+	}
+
+	if _, err := os.Stat(sifImage); os.IsNotExist(err) {
+		sylog.Fatalf("Failed to determine image absolute path for %s: %s", sifImage, err)
+	}
+	if _, err := os.Stat(bzImage); os.IsNotExist(err) {
+		sylog.Fatalf("This functionality is not supported")
+	}
+	if _, err := os.Stat(initramfs); os.IsNotExist(err) {
+		sylog.Fatalf("This functionality is not supported")
+	}
+
+	args[0] = "Sylabs"
+	cmd := exec.Command(hvExec)
+	cmd.Args = args
+	cmd.Env = os.Environ()
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	if VMErr || debug {
+		cmd.Stderr = os.Stderr
+	}
+
+	if err := cmd.Run(); err != nil {
+		sylog.Debugf("Hypervisor exit code: %v\n", err)
+
+		if exitErr, ok := err.(*exec.ExitError); ok {
+			//Program exited with non-zero return code
+			if status, ok := exitErr.Sys().(syscall.WaitStatus); ok {
+				sylog.Debugf("Process exited with non-zero return code: %d\n", status.ExitStatus())
+			}
+		}
+	}
+
+	return nil
 }
